askDjango/travel_blog/requirements.txt
appnope==0.1.0
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==2.0
django-allauth==0.31.0
django-appconf==1.0.2
django-debug-toolbar==1.9.1
django-imagekit==4.0.2
entrypoints==0.2.3
html5lib==1.0.1
httpie==0.9.9
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.1
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
oauthlib==2.0.6
olefile==0.44
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
pyzmq==16.0.3
raven==6.4.0
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
